[shortfin] Add end-to-end cli for testing / executing prefill (#26)
Updated for what the outgoing prefill looks like. This included
validation with llama 2 that a reasonable sequent (one two three four)
correctly predicted preceding tokens.

This should support generic export / execution of gguf prefill however
there still exists issues with the newly added decoder path.

+++ sharktank/sharktank/models/llama/llama.py
@@ -41,10 +41,10 @@ class LlamaModelConfig:
     device: Optional[torch.device] = None
 
     # Dtype to use for general FP activations not otherwise configured.
-    activation_dtype: torch.dtype = torch.float32
+    activation_dtype: torch.dtype = torch.float16
 
     # Dtype to use for attention.
-    attention_dtype: torch.dtype = torch.float32
+    attention_dtype: torch.dtype = torch.float16
 
     def create_kv_cache(self) -> BaseKVCache:
         hp = self.hp
